requirements.txt
Django==1.9
hashids==1.1.0
mollie-api-python==1.1.1
psycopg2==2.6.1
